requirements.txt
# Spotify Analytics - Requirements
# Python 3.8+

# Data Processing
pandas==2.1.0
numpy==1.24.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# API & Web
requests==2.31.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.20
psycopg2-binary==2.9.7

# Utils
python-dateutil==2.8.2
